requirements.txt
python-telegram-bot==13.13
aiohttp==3.8.4
SQLAlchemy==2.0.14
schedule==1.2.0
python-dotenv==1.0.0
tmdbv3api==1.7.7
requests==2.28.2 